+++ docs/whats_new.rst
@@ -26,6 +26,11 @@ Enhancements
 - Adding :class:`braindecode.models.AttentionBaseNet`  (:gh:`509` by `Bruno Aristimunha`_ and `Martin Wimpff`_)
 
 
+- Add an integration test for all the models (:gh: 570 by `Bruno Aristimunha`_)
+
+- Adding :class:`braindecode.models.BIOT` (:gh:`428` by `Bruno Aristimunha`_)
+
+
 Bugs
 ~~~~
 - Fix annotations bug for moabb datasets with non-zero interval (:gh: 561 by `Daniel Wilson`_)